fix signedness compare (for -Werror build) (#1181)
Fixes clang warning

ompl/geometric/planners/xxl/src/XXL.cpp:356:26: error: comparison of integers of different signs: 'int' and 'size_type' (aka 'unsigned long') [-Werror,-Wsign-compare]

ompl/demos/multilevel/MultiLevelPlanningKinematicChain.cpp:68:22: error: comparison of integers of different signs: 'value_type' (aka 'int') and 'const unsigned int' [-Werror,-Wsign-compare]

+++ demos/multilevel/MultiLevelPlanningKinematicChain.cpp
@@ -65,7 +65,7 @@ PlannerPtr GetQRRT(std::vector<int> sequenceLinks, SpaceInformationPtr si)
     for (unsigned int k = 0; k < sequenceLinks.size(); k++)
     {
         auto links = sequenceLinks.at(k);
-        assert(links < numLinks);
+        assert(static_cast<unsigned int>(links) < numLinks);
 
         OMPL_INFORM("Create MultiLevel Chain with %d links.", links);
         auto spaceK(std::make_shared<KinematicChainSpace>(links, linkLength, &envs.at(links)));
